requirements.txt
streamlit
pypdf
chromadb
sentence-transformers
google-genai
python-dotenv
requests